Introduce regression tests for PBassign.py
This tests aim at checking that changes in the code do not affect
PBassign output.

To run the tests, call:
./test_regression.py

You can also install nose and run:
nose test_regression.py

This latter option is more readable as it hide most output except is the
tests fail. Nose is available at https://nose.readthedocs.org .

Note that this commit duplicates some files from the demo1 directory.
This may not be an ideal solution.