Use print as a function
One major change between python 2 and python 3 is that python 3 uses
print as a function rather than as a keyword. Since python 2.6, the
__future__ module allows python 2 scripts to use print the same way as
python 3.

Here I imported print_function from __future__ in all appropriate
modules, and replaced all keyword calls of print by function calls.

The test suites and demo 1 continue to work with python 2.7. The
test_function test suite now works with python 3.4. The test_regression
test suite and demo 1 are still broken with python 3.

demo 2 is broken but this is unrelated to the changes of this commit.
See issue #16.