Adapt the behavior of range
In python 3, the range function returns a generator rather than a list
like in python 2. Hence, the range function in python 3 behaves the same
as the xrange function in python 2.

Here, I overwrite the range function in python 2 to behave like xrange.
All the calls to xrange are replaced by calls to the overwritten range
function.

There were no call to the python 2 range function. Yet, such call may be
needed in the future. To get the behavior of python 2 range function,
and get a list rather than a generator, one should use list(range()).