Replace all occurences of .iterkeys
The dictionary iterkeys method is not available in python 3. Instead the
keys method returns a generator.

As iterating directly on the dictionary is the same as iterating on
iterkeys, I replaced all occurence of the iterkeys method by simple call
to the dictionary. This is compatible with both python 2 and python 3
and has the exact same behavior in both the versions. It is less
explicit, though.