Copy, in test_data, the files used to demo xtc
Add to the regression test reference directory the input and reference
output files used in the demo2 to show how to use a XTC file in
PBassign. These files will be used in regression tests.

The headers in the reference output of the XTC reading test should
reference the input file as it is referenced in the regression test. The
file md_traj_4.PB.fasta was created by demo2 with PBassign input files
in the working directory. In the regression test, PBassign input
files are in a child directory; this changes the expected headers in
PBassign output file. The headers were adapted accordingly.

The output file was generated by run_demo2.sh with the version of the
commit 10cb0e29a6eebfda5ef4d126562417fa221fdfc2