Add mdanalysis to the requirements
mdanalysis is an optional requirement for PBxplore. As it is used in the
tests, it should be installed in a dev environment. In addition, it
should be installed by TravisCI.

Note that unit tests are broken for python 3 since commit
a159989e78184bbfb52d5028a3d9cb41c171222d because the tests do not handle
the lack of availability of mdanalysis on python 3.