Adapt the tests for when  MDAnalysis is missing
MDAnalysis is not always available, it is notably missing on python 3,
and the tests failed when MDAnalysis was not available.

The availability of MDAnalysis is now checked and the tests behave
accordingly:
* if MDAnalysis is available, PSassign is tested for the correct output
  when inputed a XTC file
* if MDAnalysis is not available, PBassign should fail

I also removed MDAnalysis from the requirements. Instead I added a
conditional install of MDAnalysis in the travis file.