Fix demo of assignement on a directory
Demo 1 used the `-d` argument for PBassign to assign PB to all the files
in a directory. This option does not exist. Instead, the `-p` option
should be used; if the path given to PBassign with `-p` is a directory,
then PBassign will work on all PDB and mmcif files in that directory.