Move the PB_assign function to PBlib.py
Move the PB_assign function from PBassign.py to PBlib.py in order to
have it usable outside of PBxplore command line tools. This is a first
step toward the goal set by my proposal in issue #25 to have a PBxplore
library.

Moving the PB_assign function required to move several helper functions
that it calls. It also required to add some argument to the function as
it was using variables outside of its scope (file name of the outputs).

Next step is to split the PB_assign function.