Cleanup on PBassign regression tests
The regression tests for PBassign use the template class only partially
as these tests were design diferently from the tests for the other
tools. Indeed, each test for PBassign is actually a test matrix rather
than just a test. The same test is run with multiple files and multiple
file formats.

This commit makes the TestPBassign class inherit from the
TemplateTestCase anyway as it shares the SetUp and the TearDown methods.
This commit also move the function that are specific to PBassign tests
in the TestPBassign class.