PBlib.hclust does not ask for the sequences
The `hclust` function required the numpy array of sequences in addition to
the distance matrix. Only the sequence names were used, and they were
modified to follow a seqX scheme before being feed to the function. The
names were passed to `matrix_to_str`, which used them headers. These
headers were read by the `read.table` function in R to have predictable
names for sequences.

If no header is provided to `read.table`, then the column (and therefore
the sequences) are named V1, V2... So the default names are already
predictable and there is no need for custom names to be passed to
`hclust`.

The `matrix_to_str` was also used to write the distance matrix to a
file. In this case, the sequence names could be useful for human
readers.

Though, the names passed were always of the form seq0, seq1, seq2...
Thus, the names can be generated from the shape of the distance matrix
and not passed to the function. In addition, names like that are not
useful to a human reader, and a distance matrix for more than a few
sequences is unreadable anyway.

Because it was useless, `matrix_to_str` does not write the sequence names
anymore. Thus, it does not ask for them anymore. As a consequence,
`hclust` and `write_distance_matrix` do not ask for them neither. In
addition, I renamed `matrix_to_str` to `_matrix_to_str`. Indeed, this
function should not be part of the API and will most likely be discarded
when we will get rid of R for the clustering. The `_` prefix indicates
that.

I also simplified the `write_distance_matrix` to use numpy.savetxt
instead of `_matrix_to_str` that should disappear at one point.

Finally, having the `hclust` function asking only for the minimum is a
nice cleanup of the API.