Cleanup `PBclust.compare`
This commit introduces some cleanup in the `PBclust.compare` function.
This function wraps the comparison of the sequences with the first one
for the command line. Especially, it prints informations about the
process, it adapt the substitution matrix to get one digit per position,
and it writes the result in a file.

This commit adds a docstring to the `PBclust.compare` function.

This commit also change how the fasta file is written. So far, the fasta
file was written usinf the `PBlib.write_fasta` function. This function
opens the fasta file on 'append' mode, writes the fasta entry, and close
the file. This means that the output file was opened and closed once per
entry. It also means that the file had to be cleared before the first
entry is written not to stack the output of successive runs of the
program. This commit introduces the `PBlib.write_fasta_entry` that
writes a fasta entry in an already open file. Using this function, the
file is opened and closed only one. Also, the `PBlib.clean_file` that
delete the file if it exists is not needed anymore.

`PBlib.write_fasta` and 'PBlib.clean_fasta` are used on other places.
Therefore, this commit do not remove them from the library. Yet, they
are error prone function and should be considered as deprecated.

The `PBlib.write_fasta_entry` is not ideal. I may introduce latter a
`FastaWritter` class that could be used as a context manager. Maybe even
a `Fasta` class to do any operation doable on a fasta file. However,
this requires a careful design.