Cleanup `PBclust.write_clusters` and kill seq_id
This commit adds a docstring to the `PBclust.write_clusters` function,
and clean up that function.

The cleanup made clear that the seq_id list was useless. This list is
written by the R section of the `PBlib.hclust` function. It contains the
column names of the distance matrix; these names corresponds to the
sequences names. Until commit d9aaae1, the sequence names were given
following a seqX scheme; since this commit, they are not given anymore
and so use the default scheme in R that is VX. Because the
`PBlib.hclust` function striped the constant part of the sequence name,
the `seq_id` list was basically the same as range(len(sequences)).

Because the `seq_id` list was useless, this commit removes it
completely. This changes how to call `PBclust.write_clusters`, and what
`PBlib.hclust` returns.