Avoid crash of demo2 when weblogo is not available
If weblogo is not available, then PBstat exits with a non 0 code.
Because run_demo2.sh has `set -e`, it exits when it encounter the non 0
code of PBstat. Therefore, the demo do not run until the end.

This commit display a message if weblogo is not installed, and allow the
demo to continue.